+++ NEWS.md
@@ -6,11 +6,12 @@
 
 ## Changes to functions
 
-* `grpmean()` now uses `contrasts()` from package *emmeans* to compute p-values, which now correclty indicate whether the sub-group mean is significantly different from the total mean.
+* `grpmean()` now uses `contrasts()` from package *emmeans* to compute p-values, which correclty indicate whether the sub-group mean is significantly different from the total mean.
 
 ## Bug fixes
 
 * `model_frame()` did not work properly for variables that were standardized with `scale()`.
+* In certain cases, `weight.by`-argument did  not work in `grpmean()`.
 
 # sjstats 0.13.0